Day 1: Market Data 101
Implemented load_yfinance(ticker, start, end),
which checks if there is already a cached parquet
file with the corresponding data, and if not,loads
the data for a given ticker between the start and
end dates into a dataframe and caches it to
parquet. load_yfinance was tested with AAPL data
from the year 2024.